BLOCKS-219 split Share and Catroid library
