+++ .eslintignore
@@ -2,3 +2,4 @@
 dist/
 release/
 tools/
+release_catroid/
