+++ package.json
@@ -16,10 +16,11 @@
     "test:dev": "cross-env-shell NODE_ENV=development TYPE=testing \"yarn initiate && webpack-dev-server\"",
     "initiate": "yarn translate",
     "translate": "node i18n/create_json.js && node i18n/create_msg.js",
-    "clean": "rimraf dist ; rimraf release",
+    "clean": "rimraf dist ; rimraf release ; rimraf release_catroid",
     "release:dev": "cross-env-shell NODE_ENV=development \"yarn initiate && webpack-dev-server --config webpack.release.dev.config.js\"",
-    "release:build": "cross-env-shell NODE_ENV=production \"yarn initiate && webpack --config webpack.release.build.config.js\"",
-    "release:builddev": "cross-env-shell NODE_ENV=development \"yarn initiate && webpack --config webpack.release.build.config.js\"",
+    "release:build": "cross-env-shell NODE_ENV=production TARGET=share \"yarn initiate && webpack --config webpack.release.build.config.js\"",
+    "catroid:builddev": "cross-env-shell NODE_ENV=development TARGET=catroid \"yarn initiate && webpack --config webpack.release.build.config.js\"",
+    "catroid:build": "cross-env-shell NODE_ENV=production TARGET=catroid \"yarn initiate && webpack --config webpack.release.build.config.js\"",
     "lint:es": "yarn eslint --ext .js ./",
     "lint:pr": "yarn prettier -c src test",
     "lint": "yarn lint:es --max-warnings 0 && yarn lint:pr",
