+++ src/library/css/catroid.css
@@ -1,0 +1,15 @@
+html,
+body {
+  height: 100vh;
+  width: 100%;
+}
+
+#catroid-catblocks-container {
+  height: 100vh;
+  width: 100%;
+}
+
+#catroid-catblocks-container.blocklySvg {
+  height: 100vh;
+  width: 100%;
+}
