+++ src/library/css/share.css
@@ -74,29 +74,6 @@ img {
   vertical-align: middle;
 }
 
-.blocklyText {
-  fill: #fff;
-  font-family: 'Helvetica Neue', 'Segoe UI', Helvetica, sans-serif;
-  font-size: 12pt;
-  font-weight: bold;
-}
-
-.blocklyNonEditableText > rect:not(.blocklyDropdownRect),
-.blocklyEditableText > rect:not(.blocklyDropdownRect) {
-  fill: #fff;
-}
-
-.blocklyNonEditableText > text,
-.blocklyEditableText > text,
-.blocklyNonEditableText > g > text,
-.blocklyEditableText > g > text {
-  fill: #575e75;
-}
-
-.blocklyDropdownText {
-  fill: #fff !important;
-}
-
 .card-header {
   cursor: pointer;
 }
@@ -142,30 +119,7 @@ img {
   font-weight: bold;
 }
 
-.search {
-  border: none;
-}
-
 #catblocks-program-container {
   overflow-x: hidden;
   overflow-y: auto;
-}
-
-#formulaPopupContent {
-  word-wrap: break-word;
-}
-
-.catblocks-script-modifiable {
-  height: 80vh;
-  width: 100%;
-}
-
-.catblocks-script-modifiable.blocklySvg {
-  height: 80vh;
-  width: 100%;
-}
-
-.catblockls-blockly-invisible > .blocklyPath {
-  fill-opacity: 0;
-  stroke-opacity: 0;
 }
