--- src/library/js/index.js
@@ -1,2 +1,0 @@
-const CatBlocks = require('./lib');
-module.exports = CatBlocks.CatBlocks;
