+++ src/library/js/webpack_catroid.js
@@ -1,0 +1,2 @@
+const CatBlocks = require('./lib_catroid');
+module.exports = CatBlocks.CatBlocks;
