+++ src/library/js/webpack_share.js
@@ -1,0 +1,2 @@
+const CatBlocks = require('./lib_share');
+module.exports = CatBlocks.CatBlocks;
